apps/city_regulatory_service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
httpx>=0.25.0
python-multipart>=0.0.6
# City Regulatory Service Requirements

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Utilities
python-dateutil>=2.8.0
opentelemetry-sdk>=1.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI
